fix: modify some tests only to be accepted by dicoding reviewers
Honestly, I don't agree with the feedback by them. Some of test cases
don't need to be re-written in the repository or usecases for example
validation because it has been caught by type checking and covered by
dedicated unit test in the entity domain. I'm gonna revert this when
this commit (maybe next if I got the same issue) got passed by the
reviewers.